% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello-world.R
\name{hello_world}
\alias{hello_world}
\title{Hello World}
\usage{
hello_world(language)
}
\arguments{
\item{language}{Language to display "Hello World" should be ISO format}
}
\value{
NULL just a function with a side effect
}
\description{
Prints out "Hello World" in different languages
}
\examples{
hello_world("EN")
hello_world("CY")

}
